🦄 refactor(notification): 移动邮件相关配置至config/email中
