+++ backend/blog-backend/src/main/java/com/kisesaki/blog/notification/config/email/EmailConfigurationProperties.java
@@ -1,4 +1,4 @@
-package com.kisesaki.blog.notification.config;
+package com.kisesaki.blog.notification.config.email;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
